+++ CHANGELOG.md
@@ -6,6 +6,18 @@ File format follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 See cloning and downloading instructions [here](https://www.ivoyager.dev/developers/).
 
+## [v0.0.3] - UNRELEASED
+
+Developed using Godot 4.4.1.
+
+Breaks older savegame files!
+
+### Added
+
+* Add Timer functionality for time-based autosaves.
+* Add debug functions to test for unfreed procedural objects (likely from RefCounteds with unlisted references).
+* More and better doc comments.
+
 ## [v0.0.2] - 2025-03-20
 
 Developed using Godot 4.4.
@@ -28,4 +40,5 @@ Initial alpha release!
 
 This plugin replaces [ivoyager_tree_saver](https://github.com/ivoyager/ivoyager_tree_saver). It has that plugin's functionality plus save/load GUI classes taken out of [ivoyager_core](https://github.com/ivoyager/ivoyager_core).
 
+[v0.0.3]: https://github.com/ivoyager/ivoyager_save/compare/v0.0.2...HEAD
 [v0.0.2]: https://github.com/ivoyager/ivoyager_save/compare/v0.0.1...v0.0.2